Implemented Chat Container Component to store input messages in a clean format
ChatContainer.jsx:
Maps over every message in the msgList State Variable (array of objects that contain the messages) and creates a ChatMessage Component

ChatInput.jsx:
When a message gets sent from the user, it updates the msgList state variable. For lengthy messages, it scrolls all the way to the bottom in a clean transition using refs to manipulate the DOM.

ChatMessage.jsx:
Dyanmically rendering based on if the message is sent from the user or the AI chatbot. Passes down the msg object for each message that contains the id, sender (user or ai), and the text

Tailwind styles: