Formatted the Open API response neatly using Markdown --> HTML conversion
Using "markdown-it": "^14.1.0" and "@tailwindcss/typography": "^0.5.12"

I was able to format the response from the API to be neatly formatted.

On the server side (index.js), I pre-prompted the system to use markdown syntax for the response.

On the client side (src/components/chat/ChatMessage.jsx), I used MarkdownIt library, to convert the markdown response to HTML. I then use the tailwind typography library to format the HTML nicely.