Added basic functionality for Sidebar and New Chat features
The functionality for the sidebar is now implemented. The next steps is to log previous chats into the sidebar.

Transitions for the mode dropdown and sidebar were also implemented. Dropdown sliding down vertically was trickier and needed to find the max height of the container using a ref.

Basic button functionality for creating new chats just refreshes the page. This will be fixed soon.