Protected Routes & Client Side Routing for users and chat logs
In this commit, I implemented an impactful feature with client side routing.

User enters page -> directed to sign in
user signs in -> directed to chatPage http://localhost:5173/:userId
User types in a new message and initiates a new chat -> directed to
http://localhost:5173/:userId/chat/:chatId

The protected route component will re-direct the user back to the sign in page if they are not logged in.

If the user is signed in but they manipulate the chatId or userID, then they are directed back to:
http://localhost:5173/:userId/