CRUD Operations implemented with OpenAI Assistants
Major updates:

1. Separating slices for logs and messages.
MessageSlice operations are now separated from logSlice operations.

2. OpenAI Assistants are implemented with threads.
Users can now continue conversations with the chat model remembering previous messages.

3. Logs are stored in MongoDB Database and loaded in when a user is logged in.

**Conclusion**: A working implementation of a CRUD app that uses firebase authentication, MongoDB to store data for users, Redux for state management, and react router for client-side routing.

**One slight problem**: The streaming feature is no longer working after I implemented the Assistants. I am not sure why but that will be the next thing to fix.
Another thing to fix:
Stop storing the entire user log list in the state variable. Instead just store the logIds, titles, and timestamp to display on the sidebar. When a user clicks a log, fetch the message list from the database and load it on the screen.