Added in error handling and deleting threads
Feature 1:Error handling
If a message takes too long to respond (happens when user sends too many messages all at once), then they will get an error message that say to "try again later"

Future implementations of this:
Give warning messages for users that have used up all of their messages or tokens for the day

Feature 2: Threads are now deleted from the Assistant. In the application before, logs could be deleted but their threadID tied with the OpenAI Assistants API statyed. It was unclear on if this would have an effect, but just in case, I deleted the thread ID corresponding to the chat logID