Styled the Sidebar menu
Sidebar menu is now broken up into 3 main parts.

The header that contains the ChatLog title and the close button
The body containing a list of all the chat logs with the ability to scroll vertically
The footer containing the user menu that has the user avatar, the username, and a sign out button.

The header and footer are both sticky