Bug Fix: SectionsChosen adding new sections (#338)
* search query returns sorted sections by rating

* Add single section to calendar bug

Adding a single section to calendar when it is not currently in selected sections is fixed

* Sections added to calendar bug fix

IF a section is added to calendar that is not in the selected section list. we first add it to the selected section list then add it to our calendar from there.

* remove previous conditional changes

sections should now be defined as core issue was addressed